Add option to generate DQL and DDL statements
And release 0.16.0
