+++ test/__mocks__/schemaTableColumn.ts
@@ -20,8 +20,18 @@ export const schemaTableColumnMock = {
       'alembic_version',
       'version_num',
       'text',
+      'BASE TABLE',
     ],
-    ['gc', 'jwt_refresh_token', 'id', 'gc', 'jwt_refresh_token', 'id', 'text'],
+    [
+      'gc',
+      'jwt_refresh_token',
+      'id',
+      'gc',
+      'jwt_refresh_token',
+      'id',
+      'text',
+      'BASE TABLE',
+    ],
     [
       'information_schema',
       'character_sets',
@@ -30,6 +40,7 @@ export const schemaTableColumnMock = {
       'character_sets',
       'character_repertoire',
       'text',
+      'BASE TABLE',
     ],
     [
       'information_schema',
@@ -39,8 +50,28 @@ export const schemaTableColumnMock = {
       'character_sets',
       'character_set_catalog',
       'text',
+      'BASE TABLE',
     ],
-    ['new_schema', 'new_table', 'id', 'new_schema', 'new_table', 'id', 'bigint'],
+    [
+      'new_schema',
+      'new_table',
+      'id',
+      'new_schema',
+      'new_table',
+      'id',
+      'bigint',
+      'BASE TABLE',
+    ],
+    [
+      'new_schema',
+      'new_view',
+      'view_column',
+      'view_schema',
+      'test_view',
+      'version_num',
+      'text',
+      'VIEW',
+    ],
     [
       'pg_catalog',
       'pg_am',
@@ -49,8 +80,18 @@ export const schemaTableColumnMock = {
       'pg_am',
       'amhandler',
       'regproc',
+      'BASE TABLE',
     ],
-    ['pg_catalog', 'pg_am', 'amname', 'pg_catalog', 'pg_am', 'amname', 'text'],
+    [
+      'pg_catalog',
+      'pg_am',
+      'amname',
+      'pg_catalog',
+      'pg_am',
+      'amname',
+      'text',
+      'BASE TABLE',
+    ],
     [
       'sys',
       'allocations',
@@ -59,6 +100,7 @@ export const schemaTableColumnMock = {
       'allocations',
       'current_state',
       'text',
+      'BASE TABLE',
     ],
     [
       'sys',
@@ -68,6 +110,7 @@ export const schemaTableColumnMock = {
       'allocations',
       'decisions',
       'object_array',
+      'BASE TABLE',
     ],
     [
       'sys',
@@ -77,6 +120,7 @@ export const schemaTableColumnMock = {
       'allocations',
       "decisions['explanations']",
       'text_array_array',
+      'BASE TABLE',
     ],
   ],
   rowcount: 10,
